+++ src/ext_widgets/mutable_image/mutable_image.c
@@ -145,7 +145,12 @@ TK_DECL_VTABLE(mutable_image) = {.size = sizeof(mutable_image_t),
                                  .get_prop = image_base_get_prop};
 
 static ret_t mutable_image_invalidate(const timer_info_t* info) {
-  widget_invalidate_force(WIDGET(info->ctx), NULL);
+  widget_t* widget = WIDGET(info->ctx);
+  mutable_image_t* mutable_image = MUTABLE_IMAGE(widget);
+
+  if(mutable_image->need_redraw == NULL || mutable_image->need_redraw(mutable_image->need_redraw_ctx)) {
+    widget_invalidate_force(WIDGET(info->ctx), NULL);
+  }
 
   return RET_REPEAT;
 }
@@ -179,6 +184,17 @@ ret_t mutable_image_set_prepare_image(widget_t* widget, mutable_image_prepare_im
 
   mutable_image->prepare_image = prepare_image;
   mutable_image->prepare_image_ctx = prepare_image_ctx;
+
+  return RET_OK;
+}
+
+ret_t mutable_image_set_need_redraw(widget_t* widget, mutable_image_need_redraw_t need_redraw,
+                                      void* need_redraw_ctx) {
+  mutable_image_t* mutable_image = MUTABLE_IMAGE(widget);
+  return_value_if_fail(mutable_image != NULL && need_redraw != NULL, RET_BAD_PARAMS);
+
+  mutable_image->need_redraw = need_redraw;
+  mutable_image->need_redraw_ctx = need_redraw_ctx;
 
   return RET_OK;
 }